Refactor foundation tests to use the same namespace structure as projects being tested
