+++ backend-foundation/Tests/DomainModeling/Identity/IdGeneratorTests.cs
@@ -2,7 +2,7 @@ using FluentAssertions;
 using PlatformPlatform.Foundation.DomainModeling.Identity;
 using Xunit;
 
-namespace PlatformPlatform.Foundation.Tests.Domain;
+namespace PlatformPlatform.Foundation.Tests.DomainModeling.Identity;
 
 public class IdGeneratorTests
 {
