+++ backend-foundation/Tests/PersistenceInfrastructure/Persistence/UnitOfWorkTests.cs
@@ -5,7 +5,7 @@ using PlatformPlatform.Foundation.PersistenceInfrastructure.Persistence;
 using PlatformPlatform.Foundation.Tests.TestEntities;
 using Xunit;
 
-namespace PlatformPlatform.Foundation.Tests.Infrastructure;
+namespace PlatformPlatform.Foundation.Tests.PersistenceInfrastructure.Persistence;
 
 public sealed class UnitOfWorkTests : IDisposable
 {
